Make fixtures available for all portadmin tests
